refactor: componentize address reducer
