+++ renderer/reducers/address/constants.js
@@ -1,0 +1,6 @@
+export const FETCH_ADDRESSES = 'FETCH_ADDRESSES'
+export const FETCH_ADDRESSES_SUCCESS = 'FETCH_ADDRESSES_SUCCESS'
+export const FETCH_ADDRESSES_FAILURE = 'FETCH_ADDRESSES_FAILURE'
+export const NEW_ADDRESS = 'NEW_ADDRESS'
+export const NEW_ADDRESS_SUCCESS = 'NEW_ADDRESS_SUCCESS'
+export const NEW_ADDRESS_FAILURE = 'NEW_ADDRESS_FAILURE'
