+++ renderer/reducers/address/index.js
@@ -1,0 +1,6 @@
+import addressReducer from './reducer'
+
+export default addressReducer
+export addressSelectors from './selectors'
+export * from './constants'
+export * from './reducer'
